+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
-#### 5.158.1 - 2018-05-08
-BUGFIX: Fixed incorrect framework restrictions in the lockfile - https://github.com/fsprojects/Paket/pull/3176
+#### 5.159.0 - 2018-05-08
+* Allows to add git-repositories to paket.lock and paket.references files via CLI - https://github.com/fsprojects/Paket/pull/3125
+* BUGFIX: Fixed incorrect framework restrictions in the lockfile - https://github.com/fsprojects/Paket/pull/3176
 
 #### 5.158.0 - 2018-05-08
 * BUGFIX: Paket restore silently failed when TargetFramework(s) are specified in Directory.Build.props and not csproj - https://github.com/fsprojects/Paket/pull/3013